Resolved #357 - Allow usage of Parcelable for IHeaders and IFlexible
If you pass the Items or Headers to an Intent as a Parcelable the headers
are recreated as different instances and are shown above each item and not
the whole group. The reason for this of the usage of reference equality
check. We need to use the equal() to fix this issue.